Fix int overflow when the file size is too big for an int

+++ astropy/io/ascii/src/tokenizer.h
@@ -54,8 +54,8 @@ typedef enum
 typedef struct
 {
     char *source;          // single string containing all of the input
-    int source_len;        // length of the input
-    int source_pos;        // current index in source for tokenization
+    unsigned long source_len; // length of the input
+    unsigned long source_pos; // current index in source for tokenization
     char delimiter;        // delimiter character
     char comment;          // comment character
     char quotechar;        // quote character
